refactor: remove file prefix
